refactor: move producer/consumer code to prod_cons/
Update README.md to reflect new layout and intent.
