+++ src/prod_cons/consumer.ts
@@ -6,7 +6,7 @@ import {
     DocumentID,
     Subscription,
 } from '@dittolive/ditto'
-import { DEFAULT_WEBUI_PORT } from './default.js'
+import { DEFAULT_WEBUI_PORT } from '../default.js'
 import Fastify, { FastifyInstance } from 'fastify'
 import { existsSync, mkdirSync } from 'fs'
 import fastifyStatic from '@fastify/static'
